Delete User and Regex for Groups

See merge request bettercollected/backend!44

+++ backend/backend/app/controllers/auth_router.py
@@ -1,5 +1,6 @@
 """Auth controller implementation."""
 import logging
+from http import HTTPStatus
 from typing import Optional
 
 from classy_fastapi import Routable, get, post, delete
@@ -8,6 +9,8 @@ from starlette.requests import Request
 from starlette.responses import RedirectResponse, Response
 
 from backend.app.container import container
+from backend.app.exceptions import HTTPException
+from backend.app.models.dtos.user_status_dto import UserStatusDto
 from backend.app.router import router
 from backend.app.services.auth_cookie_service import (
     delete_token_cookie,
@@ -18,7 +21,10 @@ from backend.app.services.user_service import (
     get_logged_user,
     get_user_if_logged_in,
     add_refresh_token_to_blacklist,
+    get_access_token,
+    get_refresh_token,
 )
+from backend.config import settings
 from common.models.user import User, UserLoginWithOTP
 
 log = logging.getLogger(__name__)
@@ -31,7 +37,7 @@ class AuthRoutes(Routable):
         super().__init__(*args, **kwargs)
         self.auth_service = auth_service
 
-    @get("/status")
+    @get("/status", response_model=UserStatusDto)
     async def status(self, user: User = Depends(get_logged_user)):
         return await self.auth_service.get_user_status(user)
 
@@ -118,10 +124,24 @@ class AuthRoutes(Routable):
     async def delete_user(
         self,
         request: Request,
-        response: Response,
         user: User = Depends(get_logged_user),
     ):
-        await add_refresh_token_to_blacklist(request=request)
+        if request.headers.get("api_key") != settings.temporal_settings.api_key:
+            raise HTTPException(
+                status_code=HTTPStatus.FORBIDDEN,
+                content="You are not allowed to perform this action.",
+            )
         await self.auth_service.delete_user(user=user)
-        delete_token_cookie(response=response)
+        await add_refresh_token_to_blacklist(request=request)
         return "User Deleted Successfully"
+
+    @post("/user/delete/workflow")
+    async def add_workflow_to_delete_user(
+        self,
+        access_token=Depends(get_access_token),
+        refresh_token=Depends(get_refresh_token),
+        user: User = Depends(get_logged_user),
+    ):
+        return await self.auth_service.add_workflow_to_delete_user(
+            access_token=access_token, refresh_token=refresh_token, user=user
+        )
